Refactor block_split and move to coordinates module
Simplify the code by using the `n_1d_arrays` utility function.

Had to move `check_data` to `utils` to avoid a circular dependency
problem. Not a big deal since it really doesn't depend on anything in
`base`.
